requirements.txt
streamlit==1.48.1
requests==2.32.4
pandas==2.3.1
numpy==1.26.4
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.29.0
matplotlib==3.8.2
scikit-learn==1.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
